+++ src/apps/Altinn.AccessManagement/src/Altinn.AccessMgmt.Integration/Models/PackagePermission.cs
@@ -1,0 +1,19 @@
+﻿using Altinn.AccessMgmt.Core.Models;
+
+namespace Altinn.AccessMgmt.Integration.Models;
+
+/// <summary>
+/// Package permissions
+/// </summary>
+public class PackagePermission
+{
+    /// <summary>
+    /// Package the permissions are for
+    /// </summary>
+    public CompactPackage Package { get; set; }
+
+    /// <summary>
+    /// Parties with permissions
+    /// </summary>
+    public List<ConnectionPermission> Permissions { get; set; }
+}
