+++ src/apps/Altinn.AccessManagement/src/Altinn.AccessMgmt.Integration/Models/Permission.cs
@@ -1,0 +1,34 @@
+﻿using Altinn.AccessMgmt.Core.Models;
+
+namespace Altinn.AccessMgmt.Integration.Models;
+
+/// <summary>
+/// Permission
+/// </summary>
+public class Permission
+{
+    /// <summary>
+    /// From party
+    /// </summary>
+    public CompactEntity From { get; set; }
+
+    /// <summary>
+    /// To party
+    /// </summary>
+    public CompactEntity To { get; set; }
+
+    /// <summary>
+    /// Via party
+    /// </summary>
+    public CompactEntity Via { get; set; }
+
+    /// <summary>
+    /// Role
+    /// </summary>
+    public CompactRole Role { get; set; }
+
+    /// <summary>
+    /// Via role
+    /// </summary>
+    public CompactRole ViaRole { get; set; }
+}
