+++ src/apps/Altinn.AccessManagement/src/Altinn.AccessMgmt.Integration/Models/ResourcePermission.cs
@@ -1,0 +1,19 @@
+﻿using Altinn.AccessMgmt.Core.Models;
+
+namespace Altinn.AccessMgmt.Integration.Models;
+
+/// <summary>
+/// Resource permission
+/// </summary>
+public class ResourcePermission
+{
+    /// <summary>
+    /// Resource the permissions are for
+    /// </summary>
+    public CompactResource Resource { get; set; }
+
+    /// <summary>
+    /// Parties with permissions
+    /// </summary>
+    public List<ConnectionPermission> Permissions { get; set; }
+}
